lab1/requirements.txt
LIBRARIES:
    sys
    numpy
    matplotlib.pyplot